requirements.txt
reflex==0.8.17
huggingface-hub
pillow
fastapi
uvicorn
python-json-logger
pydantic
pydantic-settings
diffusers
transformers
safetensors
accelerate
torch
torchaudio
torchvision
httpx
